requirements.txt
# ScreenTranslate-AI 依赖包清单

# GUI框架
PyQt6>=6.4.0

# 截图库
mss>=6.1.0

# 图像处理
Pillow>=9.0.0

# OCR引擎及其依赖 (CPU模式)
easyocr>=1.6.0
numpy>=1.21.0,<2.0
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.5.0

# 网络请求库
httpx>=0.24.0

# 热键监听库
pynput>=1.7.0

# 打包工具
# PyInstaller>=5.0.0